fix processGrid coverage and out-of-bounds Field writes
Spiral in processGrid was capped at i<=70, leaving x in [0,30) and
[170,200) unvisited on the 200x100 grid. Derive maxRadius from the
start coord and grid dimensions instead.

Field setters/getters stored pos()'s int -1 sentinel into size_t,
making the bounds check unreachable; any out-of-bounds Coord would
silently write past the end of the vectors. Store the index as int.

+++ displacement/Field.hpp
@@ -67,7 +67,7 @@ public:
 
     void setOffset(Coord c, Point displacement)
     {
-        size_t idx = pos(c);
+        int idx = pos(c);
         if (idx < 0)
             return;
 
@@ -81,7 +81,7 @@ public:
 
     void setMatchQuality(Coord c, int count, float rms)
     {
-        size_t idx = pos(c);
+        int idx = pos(c);
         if (idx < 0)
             return;
 
@@ -91,7 +91,7 @@ public:
 
     void setMedianOffset(Coord c, Point displacement)
     {
-        size_t idx = pos(c);
+        int idx = pos(c);
         if (idx < 0)
             return;
 
@@ -102,7 +102,7 @@ public:
 
     void setBeforeCount(Coord c, float count)
     {
-        size_t idx = pos(c);
+        int idx = pos(c);
         if (idx < 0)
             return;
 
@@ -111,7 +111,7 @@ public:
 
     void setAfterCount(Coord c, float count)
     {
-        size_t idx = pos(c);
+        int idx = pos(c);
         if (idx < 0)
             return;
 
@@ -120,7 +120,7 @@ public:
 
     Point offset(Coord c)
     {
-        size_t idx = pos(c);
+        int idx = pos(c);
         if (idx < 0)
             return Point();
         return Point(m_x[idx], m_y[idx], m_z[idx]);
